NetworkLinux: fix computing the length of the packet
when using sizeof(data) previously it would use the variable data instead of the data member. This made
the sizeof use the sizeof a pointer instead of sizeof the data member. This didn't actaully seem to cause
any issues as most of the data is zeros and wouldn't influence the checksum. This fixes the length and
makes the code more genuine.